+++ modules/budget/versions.tf
@@ -18,16 +18,16 @@ terraform {
   required_version = ">= 0.13"
 
   required_providers {
-    google-beta = {
-      source  = "hashicorp/google-beta"
-      version = ">= 3.43, < 4.0"
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 3.43, < 5.0"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v11.1.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v11.2.3"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v11.1.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:budget/v11.2.3"
   }
 }