+++ modules/project_services/versions.tf
@@ -19,17 +19,17 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.43, <4.0"
+      version = ">= 3.43, < 5.0"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.43, <4.0"
+      version = ">= 3.43, < 5.0"
     }
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:project_services/v11.1.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:project_services/v11.2.3"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:project_services/v11.1.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:project_services/v11.2.3"
   }
 }