+++ modules/svpc_service_project/versions.tf
@@ -16,10 +16,20 @@
 
 terraform {
   required_version = ">=0.13.0"
+  required_providers {
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 3.50, < 5.0"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
+      version = ">= 3.50, < 5.0"
+    }
+  }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:svpc_service_project/v11.1.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:svpc_service_project/v11.2.3"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:svpc_service_project/v11.1.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:svpc_service_project/v11.2.3"
   }
 }